+++ test/karma.conf.js
@@ -36,9 +36,9 @@ module.exports = function(config) {
       {pattern: 'dist/lforms/webcomponent/*.js.map', included: false, served: true},
       'dist/lforms/webcomponent/assets/lib/zone.min.js',
       'dist/lforms/webcomponent/scripts.js',
-      'dist/lforms/webcomponent/runtime-es5.js',
-      'dist/lforms/webcomponent/polyfills-es5.js',
-      'dist/lforms/webcomponent/main-es5.js',
+      'dist/lforms/webcomponent/runtime.js',
+      'dist/lforms/webcomponent/polyfills.js',
+      'dist/lforms/webcomponent/main.js',
       'src/fhir/*/lformsFHIR.js',
       'test/karma/*.js'
     ],